Run check-tsconfig-references as part of test.sh
