+++ scripts/consumer-install.sh
