+++ test.sh
@@ -36,3 +36,5 @@ end cypress
 start schema
 ./packages/vit-s/src/schemas/schema-schema.sh
 end schema
+
+node scripts/check-tsconfig-references.mjs
